requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework
django-cors-headers
celery
redis
beautifulsoup4
openai
requests
gunicorn
openai>=1.0.0
django-cors-headers
pgvector